+++ test/crs/test_crs.py
@@ -1,6 +1,7 @@
 import concurrent.futures
 import json
 import platform
+from unittest.mock import patch
 
 import numpy
 import pytest
@@ -945,6 +946,24 @@ def test_to_wkt_none_warning(wkt_version):
     crs = CRS.from_wkt(wkt_string)
     with pytest.warns(FutureWarning, match="CRS cannot be converted to a WKT string"):
         assert crs.to_wkt(version=wkt_version) is None
+
+
+def test_to_proj4_none_warning():
+    crs = CRS("EPSG:4326")
+    with patch("pyproj.crs.crs.CRS._crs") as crs_mock, pytest.warns(
+        FutureWarning, match="CRS cannot be converted to a PROJ string"
+    ):
+        crs_mock.to_proj4.return_value = None
+        assert crs.to_proj4() is None
+
+
+def test_to_json_none_warning():
+    crs = CRS("EPSG:4326")
+    with patch("pyproj.crs.crs.CRS._crs") as crs_mock, pytest.warns(
+        FutureWarning, match="CRS cannot be converted to a PROJ JSON string"
+    ):
+        crs_mock.to_json.return_value = None
+        assert crs.to_json() is None
 
 
 def test_to_proj4_enum():